# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Commerce API Documentation",
        "description": "API support for a E-Commerce",
        "contact": {
            "email": "makinsuroju@gmail.com"
        },
        "license": {
            "name": "Apache License 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/profile": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "User profile",
                "description": "Get the authenticated user",
                "operationId": "f427d7b988216725d15b1ed89d0c7b42",
                "responses": {
                    "200": {
                        "description": "User profile"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/v1/profile-update/{id}": {
            "patch": {
                "tags": [
                    "Auth"
                ],
                "summary": "Update profile",
                "description": "Update user profile",
                "operationId": "008ad72105825eb0f5a73fdda7993628",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "204": {
                        "description": "update a user"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/v1/admin/users": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "List all users",
                "description": "List all users",
                "operationId": "f969cdd93a0254f0627c64bc21caac43",
                "responses": {
                    "200": {
                        "description": "list all users"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/v1/admin/users/{id}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "get a single user",
                "description": "get a single user",
                "operationId": "b73c9435ca17d926b15c68fcc3c044dc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get a single user"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Delete a user",
                "description": "Delete a user",
                "operationId": "4a34e52fafe30b62fbdcccfc713c6305",
                "responses": {
                    "204": {
                        "description": "delete a user"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Update a user",
                "description": "Update a user",
                "operationId": "32b377344858687317563981bceaa663",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "204": {
                        "description": "update a user"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}